Extract gesture logic to PinchZoomImage component
ImageLightbox had grown to ~340 lines mixing portal/overlay/keyboard
orchestration with pinch-zoom-pan gesture handling. Extract the gesture
surface into a self-contained component.

PinchZoomImage owns: pointer/wheel/iOS-gesture listeners, scale/tx/ty
state, pinch math, double-tap, rubber-band, transform styling. It
exposes a single onTapEmpty callback for "user tapped outside the image
while not zoomed".

ImageLightbox is now 56 lines: just the portal, overlay backdrop, close
button, Escape key, and html/body scroll lock.

Component (over action) was the right fit here: the gesture state needs
to drive inline styles and class bindings on a specific surface+image
DOM structure that we own. Actions are best for behavior layered onto
foreign elements; here we own the markup.
